reorganize Py tests to be better distributed and faster on Jenkins builds
