+++ h2o-py/tests/testdir_algos/automl/pyunit_NOPASS_automl_timing.py
@@ -1,26 +1,17 @@
 from __future__ import print_function
 import sys, os, time
 
-from h2o.exceptions import H2OTypeError
-
 sys.path.insert(1, os.path.join("..","..",".."))
-import h2o
-from tests import pyunit_utils
+from tests import pyunit_utils as pu
 from h2o.automl import H2OAutoML
 
-"""
-Those tests check time constraints on AutoML runs.
-"""
+pu.load_module("_automl_utils", os.path.join(os.path.dirname(__file__)))
+from _automl_utils import import_dataset
 
-
-def import_dataset(seed=0, larger=False):
-    df = h2o.import_file(path=pyunit_utils.locate("smalldata/prostate/{}".format("prostate_complete.csv.zip" if larger else "prostate.csv")))
-    target = "CAPSULE"
-    df[target] = df[target].asfactor()
-    #Split frames
-    fr = df.split_frame(ratios=[.8,.1], seed=seed)
-    #Set up train, validation, and test sets
-    return dict(train=fr[0], valid=fr[1], test=fr[2], target=target, target_idx=1)
+"""
+Those tests check time constraints on AutoML runs and can be fragile when run on Jenkins, 
+hence the NOPASS prefix that won't fail the build if they don't pass.
+"""
 
 
 def test_automl_stops_after_max_runtime_secs():
@@ -30,7 +21,7 @@ def test_automl_stops_after_max_runtime_secs():
     ds = import_dataset()
     aml = H2OAutoML(project_name="py_aml_max_runtime_secs", seed=1, max_runtime_secs=max_runtime_secs)
     start = time.time()
-    aml.train(y=ds['target'], training_frame=ds['train'])
+    aml.train(y=ds.target, training_frame=ds.train)
     end = time.time()
     assert abs(end-start - max_runtime_secs) < cancel_tolerance_secs, end-start
 
@@ -44,7 +35,7 @@ def test_no_model_takes_more_than_max_runtime_secs_per_model():
         aml = H2OAutoML(project_name="py_aml_max_runtime_secs_per_model_{}".format(max_runtime_secs_per_model), seed=1,
                         max_runtime_secs_per_model=max_runtime_secs_per_model,
                         max_runtime_secs=max_runtime_secs)
-        aml.train(y=ds['target'], training_frame=ds['train'])
+        aml.train(y=ds.target, training_frame=ds.train)
         models_count[max_runtime_secs_per_model] = len(aml.leaderboard)
         # print(aml.leaderboard)
     # there may be one model difference as reproducibility is not perfectly guaranteed in time-bound runs
@@ -53,7 +44,7 @@ def test_no_model_takes_more_than_max_runtime_secs_per_model():
     # TODO: add assertions about single model timing once 'automl event_log' is available on client side
 
 
-pyunit_utils.run_tests([
+pu.run_tests([
     test_automl_stops_after_max_runtime_secs,
     test_no_model_takes_more_than_max_runtime_secs_per_model,
 ])
