+++ h2o-py/tests/testdir_algos/automl/pyunit_SMOKE_automl.py
@@ -4,20 +4,24 @@ import os
 import sys
 
 sys.path.insert(1, os.path.join("..","..",".."))
-import h2o
-from tests import pyunit_utils as pu
 from h2o.automl import H2OAutoML, get_leaderboard
 from h2o.model import ModelBase
+from tests import pyunit_utils as pu
+
+pu.load_module("_automl_utils", os.path.join(os.path.dirname(__file__)))
+from _automl_utils import import_dataset
+
+"""
+Some smoke tests that should be executed on all environments/setup (for example in a setup where XGBoost is not available).
+"""
 
 
 def test_smoke_automl():
-    fr = h2o.import_file(path=pu.locate("smalldata/prostate/prostate.csv"))
-    target = "CAPSULE"
-    fr[target] = fr[target].asfactor()
     nmodels = 20  # enough models to run every step (all base models, all grids, all SEs…)
     
+    ds = import_dataset()
     aml = H2OAutoML(max_models=nmodels, project_name="py_aml_smoke", seed=42, verbosity='debug')
-    model = aml.train(y=target, training_frame=fr)
+    model = aml.train(y=ds.target, training_frame=ds.train)
     
     assert isinstance(model, ModelBase)
     lb = get_leaderboard(aml, ['algos', 'provider', 'step', 'group'])
