+++ h2o-py/tests/testdir_algos/automl/pyunit_automl_checkpoints.py
@@ -2,33 +2,23 @@ from __future__ import print_function
 import sys, os
 import tempfile
 import shutil
+
 sys.path.insert(1, os.path.join("..","..",".."))
-import h2o
-from tests import pyunit_utils
 from h2o.automl import H2OAutoML
 from h2o.automl.autoh2o import get_automl
+from tests import pyunit_utils as pu
+
+pu.load_module("_automl_utils", os.path.join(os.path.dirname(__file__)))
+from _automl_utils import import_dataset
+
 
 def automl_checkpoints():
-
-    df = h2o.import_file(path=pyunit_utils.locate("smalldata/logreg/prostate.csv"))
-
-    # Split frames
-    fr = df.split_frame(ratios=[.8,.1])
-
-    # Set up train, validation, and test sets
-    train = fr[0]
-    valid = fr[1]
-    test = fr[2]
-
-    train["CAPSULE"] = train["CAPSULE"].asfactor()
-    valid["CAPSULE"] = valid["CAPSULE"].asfactor()
-    test["CAPSULE"] = test["CAPSULE"].asfactor()
-
+    ds = import_dataset()
     checkpoints_dir = tempfile.mkdtemp()
 
     aml = H2OAutoML(project_name="py_aml0", stopping_rounds=3, stopping_tolerance=0.001, stopping_metric="AUC",
                     max_models=2, seed=1234, export_checkpoints_dir=checkpoints_dir)
-    aml.train(y="CAPSULE", training_frame=train)
+    aml.train(y=ds.target, training_frame=ds.train)
 
     get_aml = get_automl(aml.project_name)
     num_files = len([f for f in os.listdir(checkpoints_dir) if "_cv" not in f])  # do not count CV models
@@ -39,7 +29,6 @@ def automl_checkpoints():
     assert get_aml["leaderboard"].nrows == num_files, "Not all generated autoML models were saved."
 
 
-if __name__ == "__main__":
-    pyunit_utils.standalone_test(automl_checkpoints)
-else:
-    automl_checkpoints()
+pu.run_tests([
+    automl_checkpoints
+])
