+++ h2o-py/tests/testdir_algos/automl/pyunit_automl_concurrent_runs.py
@@ -1,27 +1,19 @@
-from concurrent.futures import ThreadPoolExecutor, wait
+from concurrent.futures import ThreadPoolExecutor
 from functools import partial
 from itertools import chain
 import os
 import sys
 
-from h2o.job import H2OJob
-
 sys.path.insert(1, os.path.join("..","..",".."))
-import h2o
-from tests import pyunit_utils as pu
 from h2o.automl import H2OAutoML
+from tests import pyunit_utils as pu
+
+pu.load_module("_automl_utils", os.path.join(os.path.dirname(__file__)))
+from _automl_utils import import_dataset
 
 
 def flatten(it):
     return list(chain.from_iterable(it))
-
-
-def import_dataset(seed=0):
-    df = h2o.import_file(path=pu.locate("smalldata/prostate/prostate.csv"))
-    target = "CAPSULE"
-    df[target] = df[target].asfactor()
-    fr = df.split_frame(ratios=[.8,.1], seed=seed)
-    return pu.ns(train=fr[0], valid=fr[1], test=fr[2], target=target, target_idx=1)
 
 
 def test_several_automl_instances_can_run_in_parallel():
