+++ h2o-py/tests/testdir_algos/automl/pyunit_automl_effective_parameters.py
@@ -1,11 +1,14 @@
 from __future__ import print_function
 import sys, os
+import random
+
 sys.path.insert(1, os.path.join("..","..",".."))
 import h2o
-import random
-import sys
-from tests import pyunit_utils
 from h2o.automl import H2OAutoML
+from tests import pyunit_utils as pu
+
+pu.load_module("_automl_utils", os.path.join(os.path.dirname(__file__)))
+from _automl_utils import get_partitioned_model_names
 
 # Random positive seed for AutoML
 if sys.version_info[0] < 3: #Python 2
@@ -13,17 +16,6 @@ if sys.version_info[0] < 3: #Python 2
 else: # Python 3
     automl_seed = random.randint(0, sys.maxsize)
 print("Random Seed for pyunit_automl_leaderboard.py = " + str(automl_seed))
-
-
-class Obj(object):
-    pass
-
-def get_partitioned_model_names(leaderboard):
-    model_names = Obj()
-    model_names.all = list(h2o.as_list(leaderboard['model_id'])['model_id'])
-    model_names.se = [m for m in model_names.all if m.startswith('StackedEnsemble')]
-    model_names.non_se = [m for m in model_names.all if m not in model_names.se]
-    return model_names
 
 
 def check_model_property(model_names, prop_name, present=True, actual_value=None, default_value=None, input_value=None):
@@ -42,9 +34,8 @@ def check_model_property(model_names, prop_name, present=True, actual_value=None
             assert prop_name not in model.params.keys(), "unexpected {prop} in model {model}".format(prop=prop_name, model=mn)
 
 
-
 def test_actual_default_input_stopping_rounds():
-    train = h2o.import_file(path=pyunit_utils.locate("smalldata/extdata/australia.csv"))
+    train = h2o.import_file(path=pu.locate("smalldata/extdata/australia.csv"))
     target = 'runoffnew'
     exclude_algos = ["DeepLearning", "GLM"]
     aml = H2OAutoML(project_name="actual_default_input_stopping_rounds",
@@ -53,12 +44,11 @@ def test_actual_default_input_stopping_rounds():
                     seed=automl_seed)
     aml.train(y=target, training_frame=train)
 
-    non_se = get_partitioned_model_names(aml.leaderboard).non_se
+    base_models = get_partitioned_model_names(aml.leaderboard).base
     # when using cv, all cv models are trained with the stopping_rounds = 3 (default), but the final model resets stopping_rounds to 0 and use e. g. average ntrees, iterations...
-    check_model_property(non_se, 'stopping_rounds', True, 0, 0, 3)
+    check_model_property(base_models, 'stopping_rounds', True, 0, 0, 3)
 
 
-
-pyunit_utils.run_tests([
+pu.run_tests([
     test_actual_default_input_stopping_rounds,
 ])
