+++ h2o-py/tests/testdir_algos/automl/pyunit_automl_get_automl.py
@@ -1,21 +1,21 @@
 from __future__ import print_function
 import sys, os
+
 sys.path.insert(1, os.path.join("..","..",".."))
-import h2o
-from tests import pyunit_utils
 from h2o.automl import H2OAutoML, get_leaderboard
 from h2o.automl.autoh2o import get_automl
+from tests import pyunit_utils as pu
+
+pu.load_module("_automl_utils", os.path.join(os.path.dirname(__file__)))
+from _automl_utils import import_dataset
 
 
 def test_get_automl():
-    train = h2o.import_file(path=pyunit_utils.locate("smalldata/logreg/prostate.csv"))
-    y = 'CAPSULE'
-    train[y] = train[y].asfactor()
-
+    ds = import_dataset()
     aml = H2OAutoML(project_name="test_get_automl",
                     max_models=2,
                     seed=1234)
-    aml.train(y=y, training_frame=train)
+    aml.train(y=ds.target, training_frame=ds.train)
 
     get_aml = get_automl(aml.project_name)
 
@@ -33,12 +33,15 @@ def test_get_automl():
     assert aml.training_info == get_aml.training_info
 
     # Test predictions
-    predictions = aml.predict(train)
-    predictions_from_output = get_aml.predict(train)
+    predictions = aml.predict(ds.test)
+    predictions_from_output = get_aml.predict(ds.test)
     assert (predictions == predictions_from_output).all()
 
     # Test get_leaderboard PUBDEV-7454
     assert (get_leaderboard(aml) == get_leaderboard(get_aml)).all()
     assert (get_leaderboard(aml, 'ALL') == get_leaderboard(get_aml, 'ALL')).all()
 
-pyunit_utils.run_tests([test_get_automl])
+
+pu.run_tests([
+    test_get_automl
+])
