+++ h2o-py/tests/testdir_algos/automl/pyunit_automl_memory_cleanup.py
@@ -1,19 +1,15 @@
 from __future__ import print_function
-import itertools as iter
 import sys, os
-sys.path.insert(1, os.path.join("..","..",".."))
 import re
+
+sys.path.insert(1, os.path.join("..","..",".."))
 import h2o
-from tests import pyunit_utils
+from h2o import H2OFrame
 from h2o.automl import H2OAutoML
-from h2o.frame import H2OFrame
+from tests import pyunit_utils as pu
 
-
-def get_partitioned_model_names(leaderboard):
-    model_names = [leaderboard[i, 0] for i in range(0, (leaderboard.nrows))]
-    se_model_names = [m for m in model_names if m.startswith('StackedEnsemble')]
-    non_se_model_names = [m for m in model_names if m not in se_model_names]
-    return model_names, non_se_model_names, se_model_names
+pu.load_module("_automl_utils", os.path.join(os.path.dirname(__file__)))
+from _automl_utils import import_dataset, get_partitioned_model_names
 
 
 def check_model_property(model_names, prop_name, present=True, actual_value=None, default_value=None):
@@ -69,26 +65,13 @@ def list_keys_in_memory(project_name=None):
     )
 
 
-def prepare_data(seed=1):
-    name = 'amldataset'
-    df = h2o.import_file(path=pyunit_utils.locate("smalldata/logreg/prostate.csv"), destination_frame=name)
-    target = "CAPSULE"
-    df[target] = df[target].asfactor()
-    h2o.assign(df, name)
-    fr = df.split_frame(ratios=[.8, .1],
-                        destination_frames=[name+'_'+f for f in ['training', 'validation', 'leaderboard']],
-                        seed=seed)
-    train, valid, test = fr[0], fr[1], fr[2]
-    return target, train, valid, test
-
-
 def test_suite_clean_cv_predictions():
     kcvp = 'keep_cross_validation_predictions'
     nfolds = 5
 
     def setup_and_train(param_enabled=None):
         h2o.remove_all()
-        target, train, _, _ = prepare_data()
+        ds = import_dataset()
         state = 'enabled' if param_enabled is True else 'disabled' if param_enabled is False else 'default'
         if param_enabled is None:
             aml = H2OAutoML(project_name='keep_cross_validation_predictions_'+state,
@@ -98,7 +81,7 @@ def test_suite_clean_cv_predictions():
                             nfolds=nfolds, max_models=8, seed=1,
                             keep_cross_validation_predictions=param_enabled)
 
-        aml.train(y=target, training_frame=train)
+        aml.train(y=ds.target, training_frame=ds.train)
         # print(aml.leaderboard)
         return aml
 
@@ -124,60 +107,60 @@ def test_suite_clean_cv_predictions():
     def test_default_behaviour():
         print("\n=== "+kcvp+" default behaviour ===")
         aml = setup_and_train()
-        _, non_se, se = get_partitioned_model_names(aml.leaderboard)
+        models = get_partitioned_model_names(aml.leaderboard)
         keys = list_keys_in_memory()
         preds = len(keys['cv_predictions'])
         assert preds == 0, "{preds} CV predictions were not cleaned from memory".format(preds=preds)
-        for m in non_se:
+        for m in models.base:
             assert_cv_predictions_on_model(m, False)
-        for m in se:
+        for m in models.se:
             assert not h2o.get_model(h2o.get_model(m).metalearner()['name']).cross_validation_predictions()
 
     def test_param_enabled():
         print("\n=== enabling "+kcvp+" ===")
         aml = setup_and_train(True)
-        models, non_se, se = get_partitioned_model_names(aml.leaderboard)
+        models = get_partitioned_model_names(aml.leaderboard)
         keys = list_keys_in_memory()
         preds = len(keys['cv_predictions'])
         expected = len(models) * (nfolds + 1)  # +1 for holdout prediction
         assert preds == expected, "missing CV predictions in memory, got {actual}, expected {expected}".format(actual=preds, expected=expected)
-        for m in non_se:
+        for m in models.base:
             assert_cv_predictions_on_model(m)
-        for m in se:
+        for m in models.se:
             assert_cv_predictions_on_model(h2o.get_model(m).metalearner()['name'])
 
     def test_param_disabled():
         print("\n=== disabling "+kcvp+" ===")
         aml = setup_and_train(False)
-        _, non_se, se = get_partitioned_model_names(aml.leaderboard)
+        models = get_partitioned_model_names(aml.leaderboard)
         keys = list_keys_in_memory()
         preds = len(keys['cv_predictions'])
         assert preds == 0, "{preds} CV predictions were not cleaned from memory".format(preds=preds)
-        for m in non_se:
+        for m in models.base:
             assert_cv_predictions_on_model(m, False)
-        for m in se:
+        for m in models.se:
             assert not h2o.get_model(h2o.get_model(m).metalearner()['name']).cross_validation_predictions()
 
     def test_SE_retraining_fails_when_param_disabled():
         print("\n=== disabling "+kcvp+" and retraining ===")
         total_runs = 4
         aml = setup_and_train(False)  # first run
-        _, _, first_se = get_partitioned_model_names(aml.leaderboard)
-        first_bof = next(m for m in first_se if re.search(r'_BestOfFamily_', m))
-        target, train, _, _ = prepare_data()
+        first_models = get_partitioned_model_names(aml.leaderboard)
+        first_bof = next(m for m in first_models.se if re.search(r'_BestOfFamily_', m))
+        ds = import_dataset()
         for i in range(total_runs - 1):
-            aml.train(y=target, training_frame=train)
-        _, _, se = get_partitioned_model_names(aml.leaderboard)
-        first_se_all_models = [m for m in first_se if re.search(r'_AllModels_', m)]
-        se_all_models = [m for m in se if re.search(r'_AllModels_', m)]
-        se_best_of_family = [m for m in se if re.search(r'_BestOfFamily_', m)]
+            aml.train(y=ds.target, training_frame=ds.train)
+        models = get_partitioned_model_names(aml.leaderboard)
+        first_se_all_models = [m for m in first_models.se if re.search(r'_AllModels_', m)]
+        se_all_models = [m for m in models.se if re.search(r'_AllModels_', m)]
+        se_best_of_family = [m for m in models.se if re.search(r'_BestOfFamily_', m)]
         lb = aml.leaderboard
         print(lb.head(lb.nrows))
 
-        assert len(se) == len(se_all_models) + len(se_best_of_family)
+        assert len(models.se) == len(se_all_models) + len(se_best_of_family)
         assert len(se_all_models) == len(first_se_all_models), \
             "expecting only the {} first StackedEnsemble_AllModels, but got {}".format(len(first_se_all_models), len(se_all_models))
-        assert se_all_models[0] in first_se, "first StackedEnsemble_AllModels got replaced by new one"
+        assert se_all_models[0] in first_models.se, "first StackedEnsemble_AllModels got replaced by new one"
         if len(se_best_of_family) > 1:
             assert first_bof in se_best_of_family, "first StackedEnsemble_BestOfFamily disappeared after multiple runs"
             row_of = lambda id: lb[lb['model_id'] == id]
@@ -192,14 +175,13 @@ def test_suite_clean_cv_predictions():
         print("\n=== enabling "+kcvp+" and retraining ===")
         total_runs = 4
         aml = setup_and_train(True)  # first run
-        _, _, first_se = get_partitioned_model_names(aml.leaderboard)
-        target, train, _, _ = prepare_data()
+        ds = import_dataset()
         for i in range(total_runs - 1):
-            aml.train(y=target, training_frame=train)
-        _, _, se = get_partitioned_model_names(aml.leaderboard)
-        se_all_models = [m for m in se if re.search(r'_AllModels_', m)]
-        se_best_of_family = [m for m in se if re.search(r'_BestOfFamily_', m)]
-        assert len(se) == len(se_all_models) + len(se_best_of_family)
+            aml.train(y=ds.target, training_frame=ds.train)
+        models = get_partitioned_model_names(aml.leaderboard)
+        se_all_models = [m for m in models.se if re.search(r'_AllModels_', m)]
+        se_best_of_family = [m for m in models.se if re.search(r'_BestOfFamily_', m)]
+        assert len(models.se) == len(se_all_models) + len(se_best_of_family)
         assert len(se_best_of_family) + len(se_all_models) >= total_runs, "some StackedEnsembles are missing"
 
     return [
@@ -217,7 +199,7 @@ def test_suite_clean_cv_models():
 
     def setup_and_train(param_enabled=None):
         h2o.remove_all()
-        target, train, _, _ = prepare_data()
+        ds = import_dataset()
         state = 'enabled' if param_enabled is True else 'disabled' if param_enabled is False else 'default'
         if param_enabled is None:
             aml = H2OAutoML(project_name='keep_cross_validation_models'+state,
@@ -227,59 +209,59 @@ def test_suite_clean_cv_models():
                             nfolds=nfolds, max_models=8, seed=1,
                             keep_cross_validation_models=param_enabled)
 
-        aml.train(y=target, training_frame=train)
+        aml.train(y=ds.target, training_frame=ds.train)
         # print(aml.leaderboard)
         return aml
 
     def test_default_behaviour():
         print("\n=== "+kcvm+" default behaviour ===")
         aml = setup_and_train()
-        models, non_se, se = get_partitioned_model_names(aml.leaderboard)
-        check_model_property(se, kcvm, False)
-        check_model_property(non_se, kcvm, True, False, True)
+        models = get_partitioned_model_names(aml.leaderboard)
+        check_model_property(models.se, kcvm, False)
+        check_model_property(models.base, kcvm, True, False, True)
         keys = list_keys_in_memory()
         tot, cv = len(keys['models_all']), len(keys['cv_models'])
         print("total models in memory = {tot}, among which {cv} CV models".format(tot=tot, cv=cv))
         assert tot > 0, "no models left in memory"
         assert cv == 0, "{cv} CV models were not cleaned from memory".format(cv=cv)
-        for m in non_se:
+        for m in models.base:
             assert not h2o.get_model(m).cross_validation_models(), "unexpected cv models for model "+m
-        for m in se:
+        for m in models.se:
             metal = h2o.get_model(h2o.get_model(m).metalearner()['name'])
             assert not metal.cross_validation_models(), "unexpected cv models for metalearner of model "+m
 
     def test_param_enabled():
         print("\n=== enabling "+kcvm+" ===")
         aml = setup_and_train(True)
-        models, non_se, se = get_partitioned_model_names(aml.leaderboard)
-        check_model_property(se, kcvm, False)
-        check_model_property(non_se, kcvm, True, True, True)
+        models = get_partitioned_model_names(aml.leaderboard)
+        check_model_property(models.se, kcvm, False)
+        check_model_property(models.base, kcvm, True, True, True)
         keys = list_keys_in_memory()
         tot, cv = len(keys['models_all']), len(keys['cv_models'])
         print("total models in memory = {tot}, among which {cv} CV models".format(tot=tot, cv=cv))
         assert tot > 0, "no models left in memory"
-        expected = len(models) * nfolds
+        expected = len(models.all) * nfolds
         assert cv == expected, "missing CV models in memory, got {actual}, expected {expected}".format(actual=cv, expected=expected)
-        for m in non_se:
+        for m in models.base:
             assert h2o.get_model(m).cross_validation_models(), "missing cv models for model "+m
-        for m in se:
+        for m in models.se:
             metal = h2o.get_model(h2o.get_model(m).metalearner()['name'])
             assert metal.cross_validation_models(), "missing cv models for metalearner of model "+m
 
     def test_param_disabled():
         print("\n=== disabling "+kcvm+" ===")
         aml = setup_and_train(False)
-        models, non_se, se = get_partitioned_model_names(aml.leaderboard)
-        check_model_property(se, kcvm, False)
-        check_model_property(non_se, kcvm, True, False, True)
+        models = get_partitioned_model_names(aml.leaderboard)
+        check_model_property(models.se, kcvm, False)
+        check_model_property(models.base, kcvm, True, False, True)
         keys = list_keys_in_memory()
         tot, cv = len(keys['models_all']), len(keys['cv_models'])
         print("total models in memory = {tot}, among which {cv} CV models".format(tot=tot, cv=cv))
         assert tot > 0, "no models left in memory"
         assert cv == 0, "{cv} CV models were not cleaned from memory".format(cv=cv)
-        for m in non_se:
+        for m in models.base:
             assert not h2o.get_model(m).cross_validation_models(), "unexpected cv models for model "+m
-        for m in se:
+        for m in models.se:
             metal = h2o.get_model(h2o.get_model(m).metalearner()['name'])
             assert not metal.cross_validation_models(), "unexpected cv models for metalearner of model "+m
 
@@ -303,7 +285,7 @@ def test_suite_remove_automl():
         return frame.key is not None and H2OFrame.get_frame(frame.key, rows=1) is not None
 
     def test_remove_automl_with_xval():
-        target, train, valid, test = prepare_data()
+        ds = import_dataset()
         project_name = 'aml_with_xval_remove_test'
         max_models = 5
         nfolds = 5
@@ -311,7 +293,7 @@ def test_suite_remove_automl():
                         nfolds=nfolds,
                         max_models=max_models,
                         seed=1)
-        aml.train(y=target, training_frame=train, validation_frame=valid, leaderboard_frame=test)
+        aml.train(y=ds.target, training_frame=ds.train, validation_frame=ds.valid, leaderboard_frame=ds.test)
 
         keys = list_keys_in_memory()
         assert aml.key.startswith(project_name)
@@ -343,11 +325,11 @@ def test_suite_remove_automl():
         assert len(clean['predictions']) == 0
         assert len(clean['metrics']) == 0
         assert len(clean['automl']) == 0
-        for frame in [train, valid, test]:
+        for frame in [ds.train, ds.valid, ds.test]:
             assert frame_in_cluster(frame), "frame {} has been removed from cluster".format(frame.frame_id)
 
     def test_remove_automl_with_xval_when_keeping_all_cv_details():
-        target, train, valid, test = prepare_data()
+        ds = import_dataset()
         project_name = 'aml_with_xval_remove_test'
         max_models = 5
         nfolds = 5
@@ -358,7 +340,7 @@ def test_suite_remove_automl():
                         keep_cross_validation_predictions=True,
                         keep_cross_validation_fold_assignment=True,
                         keep_cross_validation_models=True)
-        aml.train(y=target, training_frame=train)
+        aml.train(y=ds.target, training_frame=ds.train)
 
         keys = list_keys_in_memory()
         # print(keys['all'].values)
@@ -393,18 +375,18 @@ def test_suite_remove_automl():
         assert len(clean['predictions']) == 0
         assert len(clean['metrics']) == 0
         assert len(clean['automl']) == 0
-        for frame in [train, valid, test]:
+        for frame in [ds.train, ds.valid, ds.test]:
             assert frame_in_cluster(frame), "frame {} has been removed from cluster".format(frame.frame_id)
 
     def test_remove_automl_no_xval():
-        target, train, blend, test = prepare_data()
+        ds = import_dataset()
         project_name = 'aml_no_xval_remove_test'
         max_models = 5
         aml = H2OAutoML(project_name=project_name,
                         nfolds=0,
                         max_models=max_models,
                         seed=1)
-        aml.train(y=target, training_frame=train, blending_frame=blend)
+        aml.train(y=ds.target, training_frame=ds.train, blending_frame=ds.valid)
 
         keys = list_keys_in_memory()
         # print(keys['all'].values)
@@ -436,18 +418,18 @@ def test_suite_remove_automl():
         assert len(clean['metrics']) == 0
         assert len(clean['predictions']) == 0
         assert len(clean['automl']) == 0
-        for frame in [train, blend, test]:
+        for frame in [ds.train, ds.valid, ds.test]:
             assert frame_in_cluster(frame), "frame {} has been removed from cluster".format(frame.frame_id)
 
     def test_remove_automl_after_individual_manual_deletions():
-        target, train, blend, test = prepare_data()
+        ds = import_dataset()
         project_name='aml_no_xval_remove_test'
         max_models = 3
         aml = H2OAutoML(project_name=project_name,
                         nfolds=0,
                         max_models=max_models,
                         seed=1)
-        aml.train(y=target, training_frame=train, blending_frame=blend)
+        aml.train(y=ds.target, training_frame=ds.train, blending_frame=ds.blend)
 
         keys = list_keys_in_memory()
         # manually remove the first item for each category to verify robustness of global automl deletion
@@ -469,7 +451,7 @@ def test_suite_remove_automl():
         assert len(clean['metrics']) == 0
         assert len(clean['predictions']) == 0
         assert len(clean['automl']) == 0
-        for frame in [train, blend, test]:
+        for frame in [ds.train, ds.valid, ds.test]:
             assert frame_in_cluster(frame), "frame {} has been removed from cluster".format(frame.frame_id)
 
     return [
@@ -480,7 +462,7 @@ def test_suite_remove_automl():
     ]
 
 
-pyunit_utils.run_tests([
+pu.run_tests([
     test_suite_clean_cv_predictions(),
     test_suite_clean_cv_models(),
     test_suite_remove_automl()
