+++ h2o-py/tests/testdir_algos/automl/pyunit_automl_regression.py
@@ -1,33 +1,31 @@
 from __future__ import print_function
+from __future__ import print_function
 import sys, os
+
 sys.path.insert(1, os.path.join("..","..","..",".."))
 import h2o
-from tests import pyunit_utils as pu
 from h2o.automl import H2OAutoML
+from tests import pyunit_utils as pu
+
+pu.load_module("_automl_utils", os.path.join(os.path.dirname(__file__)))
+from _automl_utils import import_dataset
 
 
-def import_dataset():
-    df = h2o.import_file(path=pu.locate("smalldata/extdata/australia.csv"))
-    fr = df.split_frame(ratios=[.8,.1])
-    target = "runoffnew"
-    return pu.ns(target=target, train=fr[0], valid=fr[1], test=fr[2])
-    
-    
-def australia_automl():
-    ds = import_dataset()
-    aml = H2OAutoML(max_models=2, stopping_rounds=3, stopping_tolerance=0.001)
+def test_default_automl_with_regression_task():
+    ds = import_dataset('regression')
+    aml = H2OAutoML(max_models=2,
+                    project_name='aml_regression')
 
-    print("AutoML (Regression) run with x not provided with train, valid, and test")
-    aml.train(y=ds.target, training_frame=ds.train,validation_frame=ds.valid, leaderboard_frame=ds.test)
+    aml.train(y=ds.target, training_frame=ds.train, validation_frame=ds.valid, leaderboard_frame=ds.test)
     print(aml.leader)
     print(aml.leaderboard)
-    assert set(aml.leaderboard.columns) == set(["model_id", "mean_residual_deviance","rmse", "mse", "mae", "rmsle"])
+    assert aml.leaderboard.columns == ["model_id", "mean_residual_deviance", "rmse", "mse", "mae", "rmsle"]
 
 
 def test_workaround_for_distribution():
     try:
         h2o.rapids("(setproperty \"{}\" \"{}\")".format("sys.ai.h2o.automl.algo_parameters.all.enabled", "true"))
-        ds = import_dataset()
+        ds = import_dataset('regression')
         aml = H2OAutoML(project_name="py_test",
                         algo_parameters=dict(
                             distribution='poisson',
@@ -47,6 +45,6 @@ def test_workaround_for_distribution():
 
 
 pu.run_tests([
-    australia_automl,
+    test_default_automl_with_regression_task,
     test_workaround_for_distribution,
 ])
